feat: update Profile Screen (#347)
I updated the Profile Screen to take into account the new friendship and friend request mechanisms. It allows a user to manage their status relative to the user whose profile they get access to, directly from that screen. The friends count and animals counts are not hardcoded anymore as they are now fetched from the `UserAnimalsRepository` and the `UserFriendsRepository`.
I also improved a bit the UI, as some minor improvements were needed. These improvements include changing the background and border of the achievements preview, as well as the position of the "View all achievements" button in this preview, and improving the colours of the screen to make it more alive while keeping it sober.

## How to test
- Run `ProfileScreenTest`
- Run `ProfileScreenViewModelTest`