feat: offline disable refresh
## Description
This PR implements the disabling of the refresh functionality for the
offline mode. This does not concern exactly all the screens with a
`PullToRefreshBox` as some of them already display the `OfflineScreen`
on loss of connectivity.
This PR also introduces quite an important refactor due to having
`ConnectionManager$TooManyRequestsException`. To solve this problem I
found that defining a single `ConnectivityObserver` in
`LocalConnectivityObserver` allows me to check for connectivity anywhere
without having to define a new `ConnectivityObserver` or
`DefaultConnectivityObserver`. However this implies adding a
`CompositionLocalProvider` to the `MainActivity` and to many tests,
which explains the consequent amount of lines changed. I could not seem
to find a meaningful way to divide this PR into smaller ones as the
tests would fail and the CI would not be able to pass.

## Related issues
Closes #369 

## Trade-offs or known limitations 
- The offline mode is still not fully implemented
- From now on, we have to add to new tests the
`CompositionLocalProvider` in order for it to work

## How to test
1. Open the app with internet connected and navigate, try refreshing
screens.
2. Turn off Wi-Fi and mobile data.
3. Verify that previously seen screens can not be refreshed anymore. You
should get a `Toast` error indicating that you can not refresh while
offline.
4. For some screens, when offline you should see the `OfflineScreen`
5. You can also test by running the different test classes.