feat: add post location preview on home screen (#373)
I implemented a post location preview for posts on the Home screen.
For posts which have a defined location, the user can swipe to a static
map preview of the post's location. On a click on the map, the user is
still redirected to the Post Details screen, as when clicking on the
post's image. The map preview also includes a general location
description which is supposed to contain the country of the post, as
well as the region or city. For posts with a location, a little slider
state component is placed underneath the slider to notify the user that
they can swipe the image. It also shows on which part of the slideshow
the user is. For posts without location, the slider state component
isn't displayed and the image isn't swipeable.

## Trade-offs or known limitations
- The screen recording was taken on the emulator of my laptop which has
some issues with running smoothly, thus this is the reason why it is a
bit laggy.