feat: improve bottom bar design (#380)
I improved the design of the `BottomNavigationMenu` by changing
the icons to smoother versions, particularly for the home and report
icons which were too sharp-edged. I also changed the appearance of
unselected icons which now appear outlined, as opposed to filled before.
The collection screen's icon was also changed because the old version
wasn't representative enough of the screen's content.