fix: fix navigation while state is loading (#386)
I fixed a navigation design flaw in some top level screens, the
Home screen, the Collection screen and the Report screen. The flaw
resided in the fact that the top bar of these screens allows in theory
to navigate to the current user's profile page but the navigation was
only possible after loading the UI state from the view model, which
didn't really make sense. For example, on the Home screen, you had to
wait for the posts to load before you could navigate to the current
user's profile page without an error occurring.

I deleted the `HomeTopBar` and the `ReportScreenTopBar`, and I changed
the `CollectionScreenTopBar`. They were replaced by a modular version
`TopLevelTopBar`.

I added a screen parameter `onCurrentProfilePictureClick` to the Home
screen and the Report screen which is passed to the `TopLevelTopBar` for
redirection to the current user's profile page.

I modified the ViewModel of the three screens to load the user faster
and independently of the rest of the screen's state. This allows for
faster current user profile picture loading.