feat: implement UserRepository caching
## Description
This PR implements caching for the `UserRepositoryFirestore`.
Since the `User` object is much more complex than the user settins, I
implemented the caching using `Protocol Buffers` from `DataStore`.
To avoid using stale values, I consider values stored in the cache to be
stale after 10 minutes, after which we read from Firestore instead of
reading from the cache. However, if the user is offline, in order to
maintain functionality, we keep reading from the cache, as reading from
Firestore will not be successful.
These changes include:
- Update the `build.gradle.kts` and `libs.versions.toml` to include
`protobuf`
- Define `UserProto.proto` to describe the cached `User` data
- Introduce the interface `IUserCache` and the concrete `UserCache`
implementation used in `UserRepositoryFirestore`
- Update the `RepositoryProvider` to fit the new signature of
`UserRepositoryFirestore`
- Add a `refreshCache` funtion to the `UserRepository` to allow us to
clear the cache when refreshing a screen (`PullToRefresh`)
- Add a `FakeUserCache` for testing purpose
- Update existing test suites to verify that caching works as expected.
- Refactor test classes to fit the changes in `UserRepositoryFirestore`

## Related issues
Closes #393 

## How to test
- You can run the `UserRepositoryFirestoreTest`