feat: add post and report uploading animation (#397)
I implemented an animation to display during post and report
uploading. This animation consists of infinite waves going up the screen
while alternating color and some animal images navigating up the screen
as if they were going to the cloud. I tried to symbolise the uploading
by the fact of going up the screen. There is also a text that changes
whether the animation is displayed during a post upload or a report
upload.

## Trade-offs or known limitations
- I deleted two tests in the camera screen tests because of the new
animations. I got an exception because of the screen being always
recomposed and never in idle state which seems to be a requirement for
asserting display of composables.
- There are no tests because it is not interesting to test the animation
itself and it would be time spent for nothing as we do not really care
about testing the content of the animation.
- There are some sonar cloud issues with using
`LocalConfiguration.current.screenWidthDp` and
`LocalConfiguration.current.screenHeightDp` but the recommended fix with
`LocalWindow` is very unsatisfactory. The current version returns
accurate screen sizes while the other does not, which is problematic
since I use percentage positioning in the animation. Thus, I believe
keeping the code this way is better.

> [!NOTE]
> I used AI coding assistants to get help to learn how to animate in
Compose. Although they were pretty bad at it, I still took inspiration
from their given code.