feat: implement report repository caching
## Description
This PR implements the caching of the `ReportReporsitoryFirestore` using
`Protocol Buffers` from `DataStore` following the same idea as #396.
To avoid using stale values, I consider values stored in the cache to be
stale after 10 minutes, after which we read from `Firestore` instead of
reading from the cache. However, if the user is offline, in order to
maintain functionality, we keep reading from the cache, as reading from
`Firestore` will not be successful.
Also, as soon as a user refreshes the screen and is online, the cache is
cleared, allowing to read the values actually stored in `Firestore`.

This PR includes the following changes:
- Define `ReportProto.proto` to describe the cached `Report` data
- Introduce the interface `IReportCache` and the concrete `ReportCache`
implementation used in `ReportRepositoryFirestore`
- Update the `RepositoryProvider` to fit the new signature of
`ReportRepositoryFirestore`
- Add a `refreshCache` function to the `ReportRepository` to allow us to
clear the cache when refreshing a screen (`PullToRefresh`)
- Integrate the `refrechCache` function to all viewmodels using the
`ReportRepository`.
- Add a `FakeReportCache` for testing purposes
- Update existing test suites to verify that caching works as expected.
- Refactor test classes to fit the changes in
`ReportRepositoryFirestore`

## Related issue
Closes #406 

## How to test
- Run `ReportCacheTest` and `ReportRepositoryFirestoreTest`.

---------

Co-authored-by: leolmr1110 <leonardlemaire2006@gmail.com>