feat: cache posts (#411)
The goal of this commit is to cache posts using` Protocol Buffers` with
`DataStore` and make sure the cache works in the APK.
To avoid stale data, like in the existing user cache, cached values are
treated as invalid after 10 minutes. After that, posts are loaded from
Firestore instead of the cache. If the user is offline, we still read
from the cache so the app keeps working, since Firestore cannot be
reached.