feat: implement animal repository caching
## Description
This PR implements the caching of the `AnimalReporsitoryFirestore` and
`UserAnimalsRepositoryFirestore` using `Protocol Buffers` from
`DataStore` following the same idea as #396 and #410.
To avoid using stale values, I consider values stored in the cache to be
stale after 10 minutes, after which we read from `Firestore` instead of
reading from the cache. Which means I had to cache the
`CollectionScreen` such as to make it a refreshable screen. However, if
the user is offline, in order to maintain functionality, we keep reading
from the cache, as reading from `Firestore` will not be successful.
Also, as soon as a user refreshes the screen and is online, the cache is
cleared, allowing to read the values actually stored in `Firestore`.

This PR includes the following changes:
- Define `AnimalProto.proto` and `UserAnimalsProto` to describe the
cached `Animal` and `UserAnimals` data
- Introduce the interfaces `IAnimalCache` and `IUserAnimalsCache` and
the concrete `AnimalCache` and `UserAnimals` implementations used in
`AnimalRepositoryFirestore` and `UserAnimalsRepositoryFirestore`
- Update the `RepositoryProvider` to fit the new signatures of
`AnimalRepositoryFirestore` and `UserAnimalsRepositoryFirestore`
- Add a `refreshCache` function to the `AnimalRepositoryFirestore` and
`UserAnimalsRepositoryFirestore` to allow us to clear the cache when
refreshing a screen (`PullToRefresh`)
- Integrate the `refrechCache` function to all viewmodels using the
`AnimalRepositoryFirestore` and `UserAnimalsRepositoryFirestore`.
- Add `FakeAnimalCache` and `FakeUserAnimalsCache` for testing purposes
- Update existing test suites to verify that caching works as expected.
- Refactor test classes to fit the changes in
`AnimalRepositoryFirestore` and `UserAnimalsRepositoryFirestore`.
- Hide the `CommentInput` in `PostDetailsScreen` when offline.

## Related issue
Closes #408 

## How to test
Navigate through the collection screen, make sure the animal collections
are displayed both offline and online and that you can access the
animals details both offline and online
You can as always run the various test suites.