feat: Add Filters to HomeScreen (#391)
## Description
This PR adds the interactions with the filters from the HomeScreen UI.
Basically, there's a floating action button at the bottom right of the
screen that, when clicked, opens a Filters Manager which allows the user
to choose which filters to activate and configure them.

Here's the list of the possible filters:      
- **fromAuthor**: the user enters the name of a specific author to see
its posts, null by default
- **fromPlace**: the user enters the name of a specific location to see
posts that were taken there, null by default
- **ofAnimal**: the user enters the name of a specific animal to see any
posts with this animal appearing, null by default
- **onlyFriendsPosts**: the user activates it to only see posts from
theirs friends, false by default
- **onlyMyPosts**: the user activates it to only see its posts, false by
default

A few subtleties about these filters:
- For the **fromAuthor**, **fromPlace** and **ofAnimal** filters, once
it's set, a cross button appears next to them to individually reset
their value.
- The **onlyFriendsPosts** and **onlyMyPosts** filters can't be active
at the same time. If the user tries to enable one while the other is
already enabled, the other will automatically be deactivated.
- Filters are applied only when the user clicks the **Apply** button. If
not, they are immediately restored to their previous values. The same
happens if the user quits the Filters Manager.
- All filters can be restored to their default values by clicking on the
**Reset** button.

About the filtering system:
- For true/false filters, the logic is basic: the system checks whether
the filter is set to true. If it is, it keeps only the posts that
satisfy the corresponding condition (i.e the author is a friend for
onlyFriendsPosts, or the author is the current user for onlyMyPosts).
- For string filters, the process is more complex. In the first version,
the system only kept posts whose parameter exactly matched the filters
value, which was too strict. To make the system more flexible, the
current version performs a softer comparison:
- For the **fromPlace** and **ofAnimal** filters, it keeps posts whose
parameter is close to the filter value. Case, spaces and accents are
ignored, and partial matches are accepted.
- For the **fromAuthor** filter, it still checks for a specific user but
now it uses the `UserIndex` algorithm to find and propose a known user.
Potential users appear under the filter textInput and can be selected by
clicking on them, like for the search bar in the FriendsScreen

<img width="175" height="369" alt="Screenshot from 2025-12-11 07-23-39"
src="https://github.com/user-attachments/assets/812a37d5-8337-440c-884b-4ad9eea31911"
/>
<img width="175" height="369" alt="Screenshot from 2025-12-11 07-24-03"
src="https://github.com/user-attachments/assets/0a1ef796-63e8-46c2-bd02-f7f1fc2be6ae"
/>
<img width="175" height="369" alt="Screenshot from 2025-12-11 07-22-23"
src="https://github.com/user-attachments/assets/276cdaf4-da9f-4efd-895f-688572cd2d6c"
/>
<img width="175" height="369" alt="Screenshot from 2025-12-11 07-22-11"
src="https://github.com/user-attachments/assets/b6d82eff-f761-492c-b95e-1a6d1fe94dbc"
/>

## Related issues
Closes #372 

## Trade-offs or known limitations 
- As mentioned in the description of the filtering system, the first
version was too strict. Even tough the current version is more flexible,
it still could be improved. For example, we could implement `PlaceIndex`
and `AnimalIndex` algorithms (similar to the `UserIndex`) so that those
filters also look for and propose known places/animals.
- Test coverage is not optimal (72.5%). I tried to add more tests
targeting the uncovered lines, but this did not improve the coverage.
Even with AI, I was unable to implement tests that would increase it.
Given that the current coverage is still reasonable and the overall
project's line coverage remains really good, I suggest we acknowledge
that this specific area does not meet our expectations and, due to time
constraints related to the M3 deliverable, we accept it as it is.

## How to test
Run the app and use the filters on the home screen to verify that they
correctly filter the displayed posts.

## Notes
AI was used to help me implement tests in `HomeScreenTest.kt`, for the
*formatString* function in `HomeScreenViewModel.kt` and general
optimisation

---------

Co-authored-by: Martin Zimmer <mzimmer@ic2-adrstd-5010.intranet.epfl.ch>
Co-authored-by: Martin Zimmer <mzimmer@ic2-adrstd-5004.intranet.epfl.ch>
Co-authored-by: Martin Zimmer <mzimmer@ic2-adrstd-5007.intranet.epfl.ch>
Co-authored-by: Martin Zimmer <mzimmer@ic2-adrstd-5009.intranet.epfl.ch>
Co-authored-by: Martin Zimmer <mzimmer@ic2-adrstd-5008.intranet.epfl.ch>
Co-authored-by: Martin Zimmer <mzimmer@ic2-adrstd-5011.intranet.epfl.ch>