Fix/cleanup notification screen (#432)
I fixed the notification screen. There were two issues that needed
to be addressed. First, the date stamp on the notifications was
incorrectly formatted for some edge cases like "1 minutes ago". I fixed
it so that the plural or singular form is chosen when convenient.
Second, when a user deleted a notification, marked all notifications as
read or deleted all notifications, the screen would be set to a loading
state while the operation ran which was annoying, especially for
deleting individual notifications. I fixed this by implementing an
optimistic version of the ViewModel's operations. Now the update is
instantaneous and can be reversed in case of an exception thrown.

## Related issues
Closes #401