feat: integrate scroll state remember (#435)
I implemented a scroll state remembrance on the Home, Collection
and Report screens. The scroll state is remembered when navigating to
and from bottom level screens like Home to Profile or Home to
PostDetails, but it is not remembered when navigating between top level
screens like Home to Map. This allows a user to scroll down to a
post/report/animal, access its details, come back to the
Home/Report/Collection screen and not have to scroll all the way down
again. The following video highlights this behaviour for the Home
Screen.

https://github.com/user-attachments/assets/70addddc-0a28-4ac4-a15e-5cd13a91f8f9