fix: set error message when like offline
## Description
This pull request introduces a fix. When liking a post while offline the
app currently crashes. The problem comes from the fact that the
`isOnline` variable is by default defined as true in the `toggleLike`
function of the `HomeScreenViewModel`.
To fix it I simply pass the actual `isOnline` variable, that checks for
connectivity.
Now, when online one can like or unlike a post, when offline, you try to
like or unlike a post, you get a toast message explaining that posts
cannot be liked offline.